requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.txt
#
amqp==5.0.6
    # via
    #   -r requirements.txt
    #   kombu
asgiref==3.4.1
    # via
    #   -r requirements.txt
    #   django
astroid==2.8.2
    # via
    #   -r requirements.txt
    #   pylint
billiard==3.6.4.0
    # via
    #   -r requirements.txt
    #   celery
black==21.9b0
    # via -r requirements.txt
celery==5.1.2
    # via -r requirements.txt
click==7.1.2
    # via
    #   -r requirements.txt
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via
    #   -r requirements.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r requirements.txt
    #   celery
django==3.2.8
    # via
    #   -r requirements.txt
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
django-cors-headers==3.10.0
    # via -r requirements.txt
django-extensions==3.1.3
    # via -r requirements.txt
djangorestframework==3.12.4
    # via -r requirements.txt
gunicorn==20.1.0
    # via -r requirements.txt
isort==5.9.3
    # via
    #   -r requirements.txt
    #   pylint
kombu==5.1.0
    # via
    #   -r requirements.txt
    #   celery
lazy-object-proxy==1.6.0
    # via
    #   -r requirements.txt
    #   astroid
mccabe==0.6.1
    # via
    #   -r requirements.txt
    #   pylint
mypy-extensions==0.4.3
    # via
    #   -r requirements.txt
    #   black
pathspec==0.9.0
    # via
    #   -r requirements.txt
    #   black
platformdirs==2.4.0
    # via
    #   -r requirements.txt
    #   black
    #   pylint
prompt-toolkit==3.0.20
    # via
    #   -r requirements.txt
    #   click-repl
psycopg2-binary==2.9.1
    # via -r requirements.txt
pyjwt==2.2.0
    # via -r requirements.txt
pylint==2.11.1
    # via
    #   -r requirements.txt
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.4.4
    # via -r requirements.txt
pylint-plugin-utils==0.6
    # via
    #   -r requirements.txt
    #   pylint-django
pytz==2021.3
    # via
    #   -r requirements.txt
    #   celery
    #   django
redis==3.5.3
    # via -r requirements.txt
regex==2021.9.30
    # via
    #   -r requirements.txt
    #   black
six==1.16.0
    # via
    #   -r requirements.txt
    #   click-repl
sqlparse==0.4.2
    # via
    #   -r requirements.txt
    #   django
toml==0.10.2
    # via
    #   -r requirements.txt
    #   pylint
tomli==1.2.1
    # via
    #   -r requirements.txt
    #   black
typing-extensions==3.10.0.2
    # via
    #   -r requirements.txt
    #   astroid
    #   black
    #   pylint
vine==5.0.0
    # via
    #   -r requirements.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via
    #   -r requirements.txt
    #   prompt-toolkit
wrapt==1.12.1
    # via
    #   -r requirements.txt
    #   astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
